feat: move auth pages to main layout
