--- app/(auth)/layout.tsx
@@ -1,7 +1,0 @@
-export default function AuthLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <div className="flex w-full">{children}</div>;
-}
